cleaned up LGSSM typed version, now with moment and info form
