+++ dynamax/linear_gaussian_ssm/lgssm_test.py
@@ -20,6 +20,6 @@ def test_sample_and_fit(cls, kwargs, covariates):
     key1, key2 = jr.split(jr.PRNGKey(0))
     params, param_props = model.initialize(key1)
     states, emissions = model.sample(params, key2, num_timesteps=NUM_TIMESTEPS, covariates=covariates)
-    fitted_params, lps = model.fit_em(params, param_props, emissions, covariates=covariates, num_iters=10)
+    fitted_params, lps = model.fit_em(params, param_props, emissions, covariates=covariates, num_iters=3)
     assert monotonically_increasing(lps)
-    fitted_params, lps = model.fit_sgd(params, param_props, emissions, covariates=covariates, num_epochs=10)
+    fitted_params, lps = model.fit_sgd(params, param_props, emissions, covariates=covariates, num_epochs=3)
